m32 - ldexpf frexpf in asm

+++ libsrc/_DEVELOPMENT/math/float/math32/c/m32_logf.c
@@ -5,7 +5,7 @@ extern float m32_coeff_log[];
 
 float m32_logf (float x) __z88dk_fastcall
 {
-    int exp;
+    int8_t exp;
 
     /* zero or -ve arguments are not defined */
 
@@ -14,6 +14,6 @@ float m32_logf (float x) __z88dk_fastcall
     x = m32_frexpf(x, &exp) * 2.0 - 1.0;
     exp--;
     x = m32_polyf(x, m32_coeff_log, 8);
-    return x + 0.69314718055995 * exp;
+    return (x + 0.69314718055995 * (float)exp);
 }
